back/requirements.txt
# Supabase and PostgreSQL
supabase==2.7.4
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.7.1
python-dateutil==2.9.0

# HTTP and API clients
requests==2.32.3
httpx==0.27.0

# Authentication and JWT
pyjwt==2.8.0
cryptography>=42.0.0,<46.0.0

# Database migrations (if needed)
alembic==1.13.1

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-mock==3.14.0

# Development tools
black==24.4.2
mypy==1.10.0
pylint==3.2.2

# Utilities
python-dotenv==1.0.1
pytz==2024.1